If backupdir key is missing, use root dir to mounted disk
The user can hit enter when creating a new directory (no directory is
created). In doing so "backupdir" is set to nothing. Snapper does not
keep that key. Upon a search of the latest incremental backup we didn't
check for backupdir, we just had taken the first key (which ended up
being the UUID ). We now check for the backupdir key, and if it is
missing that means use the root directory on the mounted disk.

Fixes #13.